requirements.txt
aiohttp
apscheduler
pytz
python-chess
cairosvg
tqdm
